changes: moved all guards to shared directory
